requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv
motor
pymongo
requests
apscheduler